refactor: actions 함수 분리
