+++ app/_actions/getCompData.ts
