+++ app/_actions/getCompDetail.ts
