+++ components/module/CompListSection/CompList/CompList.tsx
@@ -5,7 +5,7 @@ import { VariantProps, cva } from 'class-variance-authority';
 import { cn } from '@/lib/utils/cn';
 
 import { Competition, CompetitionProps } from '@/@types/competition';
-import { getCompData } from '@/app/competition/getCompData';
+import { getCompData } from '@/app/_actions/getCompData';
 
 export const CompListVariants = cva(
   `
