+++ lib/hook/useSignUpForm.ts
@@ -3,12 +3,12 @@ import { useRouter } from 'next/navigation';
 import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { editSchema, signUpSchema } from '@/lib/utils/validation';
-import { editUser } from '@/app/mypage/edit/editUser';
-import { signUpUser } from '@/app/signup/signUpUser';
 import { useFormState } from 'react-dom';
 import type { SignUpFormValues, SignUpState } from '@/@types/signup';
 import type { FieldPath } from 'react-hook-form';
 import type { UserData } from '@/@types/user';
+import { editUser } from '@/app/_actions/editUser';
+import { signUpUser } from '@/app/_actions/signUpUser';
 
 const useSignUpForm = (userData: UserData | undefined) => {
   const router = useRouter();
